project_flowers/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
Pillow==10.0.0
numpy==1.24.3
tensorflow-macos==2.16.2
tensorflow-metal==0.8.0
